Add null checks for all commands
